+++ R/other_software.R
@@ -113,15 +113,9 @@ micro_works <- {
   conda activate metawrap
 
   mkdir -p result/metawrap
-  mkdir -p result/metawrap/gene_fa
-  ~/miniconda3/envs/waste/bin/prodigal -i result/megahit/contigs/${sample}.contigs.fa \\
-      -d result/prodigal/gene_fa/${sample}.gene.fa \\
-      -o result/prodigal/gene_gff/${sample}.gene.gff \\
-      -p meta -f gff
-
-  mkdir -p result/prodigal/fullgene_fa
-  grep 'partial=00' result/prodigal/gene_fa/${sample}.gene.fa | cut -f1 -d ' '| sed 's/>//' > result/prodigal/gene_fa/${sample}.fullid
-  seqkit grep -f result/prodigal/gene_fa/${sample}.fullid result/prodigal/gene_fa/${sample}.gene.fa > result/prodigal/fullgene_fa/${sample}.gene.fa
+  metawrap binning -o result/binning/INITIAL_BINNING_${sample} \\
+    -t 8 -m 8 -a result/megahit/contigs/${sample}.contigs.fa \\
+     --metabat2 --maxbin2 --concoct data/rm_human/${sample}.*.fq
 
 "),
     "cluster" = paste0('